# [Enhancement] Live preview speedup