--- a/src/seedsigner/gui/screens/scan_screens.py
+++ b/src/seedsigner/gui/screens/scan_screens.py
@@ -65,6 +65,8 @@ class ScanScreen(BaseScreen):
             from timeit import default_timer as timer
 
             instructions_font = Fonts.get_font(GUIConstants.BODY_FONT_NAME, GUIConstants.BUTTON_FONT_SIZE)
+
+            framerate = 0
             while self.keep_running:
                 start = timer()
                 frame = self.camera.read_video_stream(as_image=True)
@@ -84,13 +86,14 @@ class ScanScreen(BaseScreen):
                             (self.render_rect[0], self.render_rect[1])
                         )
 
-                        if scan_text:
+                        if scan_text or True:
                             self.renderer.draw.text(
                                 xy=(
                                     int(self.renderer.canvas_width/2),
                                     self.renderer.canvas_height - GUIConstants.EDGE_PADDING
                                 ),
-                                text=scan_text,
+                                # text=scan_text,
+                                text=f"{framerate:0.2f} fps",
                                 fill=GUIConstants.BODY_FONT_COLOR,
                                 font=instructions_font,
                                 stroke_width=4,
@@ -101,7 +104,8 @@ class ScanScreen(BaseScreen):
                         self.renderer.show_image()
 
                         end = timer()
-                        print(f"{1.0/(end - start):0.2f} fps") # Time in seconds, e.g. 5.38091952400282
+                        framerate = 1.0/(end - start)
+                        # print(f"{framerate:0.2f} fps") # Time in seconds, e.g. 5.38091952400282
 
                 if self.camera._video_stream is None:
                     break