--- a/src/seedsigner/hardware/camera.py
+++ b/src/seedsigner/hardware/camera.py
@@ -1,5 +1,4 @@
 import io
-import numpy
 
 from picamera import PiCamera
 from PIL import Image