Due do slower rendering times in SeedSigner OS (full-screen updates are noticeably slower), we need to optimize the `ScanScreen` live preview as much as possible. Current `dev` branch is only rendering the live preview at about 1.6 fps for the live display and 3.1 fps for the QR decoder in SeedSigner OS (see my [benchmarking branch](https://github.com/kdmukai/seedsigner/tree/dev_livepreview_benchmarking) to get an onscreen fps display).

In my testing, this PR gets a Pi Zero running SeedSigner OS up to about 3.3 fps for the live preview and a coincidental 3.3 fps for the QR decoder.

On Raspi OS, I've seen this same code range from 4.6fps display / 3.8 fps decode to 3.5 fps display / 4.6 fps decode.