@jdlcdl 2023-07-25
On my pi2...

Baseline w/o this pr.
RaspiOS: 4.5 fps display / 10.9 fps decode
seedsigner-os: 3.2 fps display / 6.5 fps decode

With this pr:
RaspiOS: 8.4 fps display / 9.3 fps decode
seedsigner-os: 6.3 fps display / 6.3 fps decode

Display framerate almost double, decode framerate slightly slower.

All measures above done with your dev_livepreview_benchmarking branch merged in.
At first glance, I did sense that this pr was an improvement, but I wouldn't have guessed it was 2x.
Your benchmarking branch sure would be nice to have, w/ onscreen fps visible, in a way where we could turn it on with a hidden setting, maybe thru a SettingsQR?