@newtonick 2023-07-31
ACK Tested

I did some large PSBT scans comparing decoding performance and did not notice any difference.